+++ src/include/hosted/misc.hpp
@@ -8,6 +8,8 @@
 
 // C++ standard library
 #include <chrono>
+#include <cstddef>
+#include <ostream>
 #include <string>
 #include <thread>
 #include <utility>
@@ -25,6 +27,8 @@ inline namespace misc {
 
 extern std::string basename(const std::string& s);
 extern std::string basename(const char * const s);
+
+extern void hex_dump(std::ostream& out, const void *obj, const std::size_t size);
 
 /* ---------------------------------------------------------------------- */
 /*  */