+++ ChangeLog.md
@@ -4,6 +4,10 @@ ChangeLog
 [0.0.0.2024072601] - 2024-07-26
 -------------------------------
 
+### Added
+
+* Library: Miscellaneous: hex_dump: The first implementation.
+
 ### Changed
 
 * Library: Sleep utility: Use pass by value instead of pass by reference.